+++ .gitignore
@@ -5,3 +5,4 @@ extension
 node_modules
 artifacts
 /**/.DS_Store
+package-lock.json
