+++ package.json
@@ -34,14 +34,15 @@
   },
   "scripts": {
     "vscode:prepublish": "npm run package",
+    "compile": "webpack",
     "watch": "tsc -watch -p ./src/VSCode.Extension/tsconfig.json",
-    "package": "webpack --mode production --config ./src/VSCode.Extension/webpack.config.js"
+    "package": "webpack --mode production"
   },
   "devDependencies": {
     "@types/node": "16.x",
     "@types/vscode": "^1.73.0",
-    "typescript": "^4.8.4",
     "ts-loader": "^9.4.1",
+    "typescript": "^4.8.4",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0"
   },
