+++ webpack.config.js
@@ -1,0 +1,45 @@
+//@ts-check
+
+'use strict';
+
+const path = require('path');
+
+//@ts-check
+/** @typedef {import('webpack').Configuration} WebpackConfig **/
+
+/** @type WebpackConfig */
+const extensionConfig = {
+  target: 'node',
+	mode: 'none',
+
+  entry: './src/VSCode.Extension/extension.ts',
+  output: {
+    path: path.resolve(__dirname, 'extension'),
+    filename: 'extension.js',
+    libraryTarget: 'commonjs2'
+  },
+  externals: {
+    vscode: 'commonjs vscode' 
+  },
+  resolve: {
+    extensions: ['.ts', '.js']
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        exclude: /node_modules/,
+        use: [
+          {
+            loader: 'ts-loader'
+          }
+        ]
+      }
+    ]
+  },
+  devtool: false,
+  infrastructureLogging: {
+    level: "log",
+  },
+};
+module.exports = [ extensionConfig ];
